+++ src/phbf.c
@@ -291,9 +291,9 @@ int photorec_bf(disk_t *disk_car, partition_t *partition, const int verbose, con
 	    buffer=buffer_olddata+blocksize;
 	    if(verbose>1)
 	    {
-	      log_verbose("Reading sector %10lu/%lu\n",
-		  (unsigned long)((offset-partition->part_offset)/disk_car->sector_size),
-		  (unsigned long)((partition->part_size-1)/disk_car->sector_size));
+	      log_verbose("Reading sector %10llu/%llu\n",
+		  (unsigned long long)((offset-partition->part_offset)/disk_car->sector_size),
+		  (unsigned long long)((partition->part_size-1)/disk_car->sector_size));
 	    }
 	    disk_car->pread(disk_car, buffer, READ_SIZE, offset);
 	  }
