+++ src/phbs.c
@@ -199,7 +199,9 @@ int photorec_find_blocksize(disk_t *disk_car, partition_t *partition, const int 
       buffer=buffer_olddata+*blocksize;
       if(verbose>1)
       {
-        log_verbose("Reading sector %10lu/%lu\n",(unsigned long)((offset-partition->part_offset)/disk_car->sector_size),(unsigned long)((partition->part_size-1)/disk_car->sector_size));
+        log_verbose("Reading sector %10llu/%llu\n",
+	    (unsigned long long)((offset-partition->part_offset)/disk_car->sector_size),
+	    (unsigned long long)((partition->part_size-1)/disk_car->sector_size));
       }
       if(disk_car->pread(disk_car, buffer, READ_SIZE, offset) != READ_SIZE)
       {
